Split imgproxy parser internals
